% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicalc_functions.r
\name{QuantBD}
\alias{QuantBD}
\title{Calculate the confidence interval for a quantile 'p'}
\usage{
QuantBD(n, p, alpha = 0.05)
}
\arguments{
\item{n}{the sample size}

\item{p}{the desired quantile to be estimated from the sample}

\item{alpha}{the confidence level (default value is 0.05)}
}
\description{
Calculate the confidence interval for a quantile 'p'
}
