% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicalc_functions.r
\name{QuantCI}
\alias{QuantCI}
\title{Calculate the percentiles and confidence intervals}
\usage{
QuantCI(cfd, n, probs = seq(0.05, 0.95, 0.01), Np = NA, alpha = 0.05)
}
\arguments{
\item{cfd}{data frame providing a list grain sizes in the first variable
and the corresponding cumulative proportion finer in the second. The
grain sizes should be recorded in mm, and the proportion finer in [0,1].}

\item{n}{total number of observations upon which the cumulative
frequency distribution in \code{cfd} is based}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{Np}{if the cumulative frequency data was a sub-sample from a finite
population of size Np, then the confidence range will be smaller and a
finite population correction is applied.}

\item{alpha}{the desired confidence level for which to calculate a
confidence interval in [0,1].}
}
\description{
\code{QuantCI} is a function that uses cumulative frequency distribution
data for Wolman or grid-by-number samples of bed surface texture to
estimate the value of user-specified percentiles. The function also uses
binomial theory to estimate the confidence intervals corresponding to a
user-specified confidence level, based on the number of grain size
measurements that were used to construct the cumulative frequency
distribution.
}
\details{
The function returns a data frame listing the estimate of each quantile,
the upper limit, and the lower limit. This method is applicable to all
quantiles between the 5th and 95th percentile of the distirbution, but
not beyond that.
}
