% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicalc_functions.r
\name{PolyCI}
\alias{PolyCI}
\title{Generate a polygon representing the confidence bounds for a grain size distribution}
\usage{
PolyCI(cfd, n, probs = seq(0.05, 0.95, 0.01), Np = NA, alpha = 0.05,
  plot = FALSE)
}
\arguments{
\item{cfd}{data frame providing a list grain sizes in the first variable
and the corresponding cumulative proportion finer in the second. The
grain sizes should be recorded in mm, and the proportion finer in [0,1].}

\item{n}{total number of observations upon which the cumulative
frequency distribution in \code{cfd} is based}

\item{probs}{numeric vector of probabilities defining the polygon vertices,
with values in [0,1].}

\item{Np}{if the cumulative frequency data was a sub-sample from a finite
population of size Np, then the confidence range will be smaller and a
finite population correction is applied.}

\item{alpha}{the desired confidence level for which to calculate a
confidence interval in [0,1].}

\item{plot}{optional flag to produce a graph of the resulting grain size
distribution}
}
\description{
\code{PolyCI} is a function that generates a data frame containing the
coordinates defining a polygon representing the confidence bounds around
a bed surface grain size distribution. The function contains an option to
generate a plot of the distribution showing the estimates of the percentiles
between the D5 and the D95, as well as the confidence limits about those
estimates.
}
